+++ tests/integration/cases/record_wide_int/Clojure_combined@v1_11.clj
@@ -1,0 +1,14 @@
+(def my_data {
+    "quantity" 1000000
+    "big" 18446744073709551615
+    "ratio" 2.5
+    "label" "tag"
+    "ok" true
+})
+(def my_data {
+    "quantity" 1000000
+    "big" 18446744073709551615
+    "ratio" 2.5
+    "label" "tag"
+    "ok" true
+})