+++ tests/integration/cases/record_wide_int/CommonLisp_combined@ansi.lisp
@@ -1,0 +1,14 @@
+(defparameter *my_data* (list
+    (cons "quantity" 1000000)
+    (cons "big" 18446744073709551615)
+    (cons "ratio" 2.5)
+    (cons "label" "tag")
+    (cons "ok" t)
+))
+(setf *my_data* (list
+    (cons "quantity" 1000000)
+    (cons "big" 18446744073709551615)
+    (cons "ratio" 2.5)
+    (cons "label" "tag")
+    (cons "ok" t)
+))