+++ tests/integration/cases/record_wide_int/Nix@v2.nix
@@ -1,0 +1,7 @@
+let my_data = {
+  quantity = 1000000;
+  big = (builtins.fromJSON "18446744073709551615");
+  ratio = 2.5;
+  label = "tag";
+  ok = true;
+}; in my_data