+++ tests/integration/cases/record_wide_int/Racket@v8.rkt
@@ -1,0 +1,8 @@
+#lang racket
+(define my_data (hash
+    "quantity" 1000000
+    "big" 18446744073709551615
+    "ratio" 2.5
+    "label" "tag"
+    "ok" #t
+))