+++ tests/integration/cases/record_wide_int/Scala@v3.scala
@@ -1,0 +1,9 @@
+object Fixture_record_wide_int_Scala {
+val my_data = Map(
+    "quantity" -> 1000000,
+    "big" -> BigInt("18446744073709551615"),
+    "ratio" -> 2.5,
+    "label" -> "tag",
+    "ok" -> true,
+)
+}