+++ tests/integration/cases/record_wide_int/Tcl@v8_6.tcl
@@ -1,0 +1,7 @@
+set my_data [dict create \
+    "quantity" 1000000 \
+    "big" 18446744073709551615 \
+    "ratio" 2.5 \
+    "label" "tag" \
+    "ok" 1 \
+]